requirements.txt
ragas==0.1.9
FlagEmbedding==1.2.10
huggingface_hub==0.23.1
llama_index==0.10.49
numpy==1.19.3
peft==0.11.1
Pillow==10.3.0
safetensors==0.4.3
sentencepiece==0.2.0
streamlit==1.34.0
timm==1.0.7
tqdm==4.66.4
transformers==4.41.0
unstructured==0.14.2